requirements.txt
dj-database-url>=0.5.0
Django==3.0.7
psycopg2-binary==2.8.5
django-crispy-forms==1.9.1
python-dotenv==0.13.0
celery==4.4.5
redis==3.5.3
